Bureaucratic_AI_Agent/backend/pyproject.toml
[project]
name = "bureaucratic-backend"
version = "0.1.0"
description = "Backend API part of the project"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.80",
    "django>=6.0.3",
    "django-jazzmin>=3.0.4",
    "django-storages>=1.14.6",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
